refactor(desktop): consolidate composer helpers
